# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Laya API",
    "description": "Run Laya typed decisions (choice / score / noul) over text, JSON objects or conversation turns, using your own questions or a built-in preset. Requests are auto-routed to the best checkpoint, or pinned with `model`.\n\nAuthenticate with an API key (`X-API-Key` or `Authorization: Bearer`) or HTTP Basic, whichever is configured on the server.",
    "version": "0.6.1"
  },
  "paths": {
...
    "/models": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "List checkpoints",
        "description": "Available checkpoints and which are currently resident.",
        "operationId": "list_models_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "503": {
            "description": "Model not loaded yet"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      }
    },
    "/presets": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "List built-in question presets",
        "description": "Return the ready-to-use question sets (triage/email/guard/moderation/router).",
        "operationId": "list_presets_presets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "object",
                  "title": "Response List Presets Presets Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "503": {
            "description": "Model not loaded yet"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      }
    },
    "/detect": {
      "post": {
        "tags": [
          "meta"
        ],
        "summary": "Detect script and language",
        "description": "Report script, best-effort language and `is_english` for a state (what routing uses).",
        "operationId": "detect_detect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Detect Detect Post"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "503": {
            "description": "Model not loaded yet"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      }
    },
    "/email/state": {
      "post": {
        "tags": [
          "meta"
        ],
        "summary": "Build a clean email state",
        "description": "Clean an email body and structure it as a state for `/predict`.",
        "operationId": "email_state_endpoint_email_state_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Email State Endpoint Email State Post"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "503": {
            "description": "Model not loaded yet"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      }
    },
    "/predict": {
      "post": {
        "tags": [
          "predict"
        ],
        "summary": "Predict one state",
        "description": "Run typed questions over a single state and return calibrated answers.",
        "operationId": "predict_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "503": {
            "description": "Model not loaded yet"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      }
    },
    "/predict/bulk": {
      "post": {
        "tags": [
          "predict"
        ],
        "summary": "Predict many states",
        "description": "Run typed questions over many states.\n\nAccepts either `states` with shared `questions`/`preset`, or `items` where each\nstate can override its questions and model. Laya's public API predicts one state\nat a time (its internals can batch, but the public `Agent.predict` cannot), so\nthis loops. Per-state errors are isolated and returned inline.",
        "operationId": "predict_bulk_predict_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkPredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkPredictResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "503": {
            "description": "Model not loaded yet"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      }
    },
    "/v1/systemone": {
      "post": {
        "tags": [
          "systemone",
          "predict"
        ],
        "summary": "SystemOne / TypeSafe-compatible prediction",
        "description": "Evaluate a SystemOne / TypeSafe-shaped request using a Laya checkpoint.",
        "operationId": "systemone_predict_v1_systemone_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemOneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemOneResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "503": {
            "description": "Model not loaded yet"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "BulkPredictRequest": {
        "properties": {
          "states": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    {
                      "items": {},
                      "type": "array"
                    }
                  ]
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "States",
            "description": "States to classify with the shared `questions`/`preset`."
          },
          "questions": {
            "anyOf": [
              {
                "additionalProperties": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ChoiceQuestion"
                    },
                    {
                      "$ref": "#/components/schemas/ScoreQuestion"
                    },
                    {
                      "$ref": "#/components/schemas/NoulQuestion"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "type",
                    "mapping": {
                      "choice": "#/components/schemas/ChoiceQuestion",
                      "noul": "#/components/schemas/NoulQuestion",
                      "score": "#/components/schemas/ScoreQuestion"
                    }
                  }
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Questions",
            "description": "Questions applied to each state."
          },
          "preset": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "triage",
                  "email",
                  "guard",
                  "moderation",
                  "router"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset",
            "description": "Preset questions for each state."
          },
          "model": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "english",
                  "multilingual",
                  "typed-decisions"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "Pin a checkpoint for all states; omit to auto-route."
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/BulkItem"
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Items",
            "description": "Per-state items, each optionally overriding questions and model."
          }
        },
        "type": "object",
        "title": "BulkPredictRequest",
        "description": "A batch with shared questions or optional per-item question and model overrides.",
        "examples": [
          {
            "questions": {
              "department": {
                "criteria": {
                  "billing": "invoices, payments, refunds",
                  "sales": "new purchases",
                  "technical": "bugs and outages"
                },
                "instructions": "Which team should handle this request?",
                "type": "choice"
              },
              "refund": {
                "instructions": "Does the customer ask for money back?",
                "type": "noul"
              },
              "urgency": {
                "criteria": [
                  "not urgent",
                  "soon",
                  "critical"
                ],
                "instructions": "How urgent is this request?",
                "type": "score"
              }
            },
            "states": [
              "I was billed twice. Please refund the duplicate today.",
              "The app crashes on launch."
            ]
          },
          {
            "items": [
              {
                "state": "I was billed twice."
              },
              {
                "model": "english",
                "state": "The app crashes."
              }
            ],
            "preset": "triage"
          }
        ]
      },
      "BulkPredictResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/BulkItemResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "count",
          "results"
        ],
        "title": "BulkPredictResponse",
        "description": "Ordered results and item count for a completed batch."
      },
...
      "ChoiceQuestion": {
        "properties": {
          "type": {
            "type": "string",
            "const": "choice",
            "title": "Type"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              }
            ],
            "title": "Instructions",
            "description": "What to decide, phrased as a question."
          },
          "criteria": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criteria",
            "description": "Option label -> description, or a plain list of labels. Descriptions may be strings or any JSON value (rendered as compact JSON).",
            "examples": [
              {
                "billing": "invoices, payments, refunds",
                "technical": [
                  "bugs",
                  "outages"
                ]
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "type",
          "instructions"
        ],
        "title": "ChoiceQuestion",
        "description": "Pick one labelled option, e.g. a routing department."
      },
      "DetectRequest": {
        "properties": {
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              }
            ],
            "title": "State"
          }
        },
        "type": "object",
        "required": [
          "state"
        ],
        "title": "DetectRequest",
        "description": "Text or structured state submitted for language detection."
      },
      "EmailStateRequest": {
        "properties": {
          "subject": {
            "type": "string",
            "title": "Subject",
            "default": ""
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "sender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender"
          },
          "clean": {
            "type": "boolean",
            "title": "Clean",
            "description": "Strip quoted history, signatures, disclaimers.",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "body"
        ],
        "title": "EmailStateRequest",
        "description": "Email content and options used to construct a prediction state."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelsResponse": {
        "properties": {
          "available": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available"
          },
          "loaded": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Loaded"
          }
        },
        "type": "object",
        "required": [
          "available",
          "loaded"
        ],
        "title": "ModelsResponse",
        "description": "Available checkpoints and the subset currently resident in memory."
      },
...
      "PredictRequest": {
        "properties": {
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              }
            ],
            "title": "State",
            "description": "Text, JSON object, or conversation turns to analyse."
          },
          "questions": {
            "anyOf": [
              {
                "additionalProperties": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ChoiceQuestion"
                    },
                    {
                      "$ref": "#/components/schemas/ScoreQuestion"
                    },
                    {
                      "$ref": "#/components/schemas/NoulQuestion"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "type",
                    "mapping": {
                      "choice": "#/components/schemas/ChoiceQuestion",
                      "noul": "#/components/schemas/NoulQuestion",
                      "score": "#/components/schemas/ScoreQuestion"
                    }
                  }
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Questions",
            "description": "Map of question id -> typed question definition."
          },
          "preset": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "triage",
                  "email",
                  "guard",
                  "moderation",
                  "router"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset",
            "description": "Built-in question set to use instead of `questions`."
          },
          "model": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "english",
                  "multilingual",
                  "typed-decisions"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "Pin a checkpoint; omit to auto-route by language."
          }
        },
        "type": "object",
        "required": [
          "state"
        ],
        "title": "PredictRequest",
        "description": "One state with either explicit typed questions or a built-in preset.",
        "examples": [
          {
            "questions": {
              "department": {
                "criteria": {
                  "billing": "invoices, payments, refunds",
                  "sales": "new purchases",
                  "technical": "bugs and outages"
                },
                "instructions": "Which team should handle this request?",
                "type": "choice"
              },
              "refund": {
                "instructions": "Does the customer ask for money back?",
                "type": "noul"
              },
              "urgency": {
                "criteria": [
                  "not urgent",
                  "soon",
                  "critical"
                ],
                "instructions": "How urgent is this request?",
                "type": "score"
              }
            },
            "state": "I was billed twice. Please refund the duplicate today."
          },
          {
            "preset": "triage",
            "state": "Hi, we were billed twice for March."
          }
        ]
      },
      "PredictResponse": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "answers": {
            "additionalProperties": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ChoiceAnswer"
                },
                {
                  "$ref": "#/components/schemas/ScoreAnswer"
                },
                {
                  "$ref": "#/components/schemas/NoulAnswer"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "choice": "#/components/schemas/ChoiceAnswer",
                  "noul": "#/components/schemas/NoulAnswer",
                  "score": "#/components/schemas/ScoreAnswer"
                }
              }
            },
            "type": "object",
            "title": "Answers"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          },
          "routing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Routing"
              },
              {
                "type": "null"
              }
            ],
            "description": "How the checkpoint was chosen (auto-route or pinned)."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "model",
          "answers",
          "usage"
        ],
        "title": "PredictResponse",
        "description": "Validated native prediction, preserving additional fields returned by Laya."
      },
...
      "SystemOneRequest": {
        "properties": {
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              }
            ],
            "title": "State",
            "description": "The content to evaluate: a string, dict, or list."
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "System One model ID, e.g. laya-english, laya-multilingual, typed-decisions."
          },
          "questions": {
            "additionalProperties": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ChoiceQuestion"
                },
                {
                  "$ref": "#/components/schemas/ScoreQuestion"
                },
                {
                  "$ref": "#/components/schemas/NoulQuestion"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "choice": "#/components/schemas/ChoiceQuestion",
                  "noul": "#/components/schemas/NoulQuestion",
                  "score": "#/components/schemas/ScoreQuestion"
                }
              }
            },
            "type": "object",
            "title": "Questions",
            "description": "Typed questions map."
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Optional session identifier."
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User",
            "description": "Optional user identifier."
          }
        },
        "type": "object",
        "required": [
          "state",
          "model",
          "questions"
        ],
        "title": "SystemOneRequest",
        "description": "SystemOne-compatible input accepting model aliases and optional client metadata."
      },
      "SystemOneResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "Laya"
          },
          "answers": {
            "additionalProperties": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/SystemOneNoulAnswer"
                },
                {
                  "$ref": "#/components/schemas/SystemOneChoiceAnswer"
                },
                {
                  "$ref": "#/components/schemas/SystemOneScoreAnswer"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "choice": "#/components/schemas/SystemOneChoiceAnswer",
                  "noul": "#/components/schemas/SystemOneNoulAnswer",
                  "score": "#/components/schemas/SystemOneScoreAnswer"
                }
              }
            },
            "type": "object",
            "title": "Answers"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "model",
          "answers",
          "usage"
        ],
        "title": "SystemOneResponse",
        "description": "SystemOne-compatible prediction with a generated request identifier."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "description": "A key from the server's API_KEYS.",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "description": "An API key sent as a bearer token.",
        "scheme": "bearer"
      },
      "BasicAuth": {
        "type": "http",
        "description": "A user:password from the server's BASIC_AUTH.",
        "scheme": "basic"
      }
    }
  },
...
}